requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
antlr4-python3-runtime=4.9.3=pypi_0
asttokens=2.2.1=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
bcrypt=4.0.1=pypi_0
blis=0.7.9=pypi_0
ca-certificates=2023.5.7=hbcca054_0
catalogue=2.0.8=pypi_0
certifi=2022.12.7=py39h06a4308_0
cffi=1.15.1=pypi_0
charset-normalizer=3.1.0=pypi_0
click=8.1.3=pypi_0
cmake=3.26.3=pypi_0
confection=0.0.4=pypi_0
contourpy=1.0.7=pypi_0
cryptography=40.0.1=pypi_0
cycler=0.11.0=pypi_0
cymem=2.0.7=pypi_0
debugpy=1.5.1=py39h295c915_0
decorator=5.1.1=pyhd8ed1ab_0
emoji=2.2.0=pypi_0
en-core-web-md=3.5.0=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
filelock=3.11.0=pypi_0
fonttools=4.39.4=pypi_0
huggingface-hub=0.13.4=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.4=pypi_0
importlib-resources=5.12.0=pypi_0
ipykernel=6.15.0=pyh210e3f2_0
ipython=8.13.2=pyh41d4057_0
ipywidgets=8.0.6=pypi_0
jedi=0.18.2=pyhd8ed1ab_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
jupyter_client=7.0.6=pyhd8ed1ab_0
jupyter_core=5.3.0=py39hf3d152e_0
jupyterlab-widgets=3.0.7=pypi_0
kiwisolver=1.4.4=pypi_0
krippendorff=0.6.0=pypi_0
langcodes=3.3.0=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
lit=16.0.1=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mpmath=1.3.0=pypi_0
murmurhash=1.0.9=pypi_0
natsort=8.3.1=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
networkx=3.1=pypi_0
numpy=1.24.2=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
omegaconf=2.3.0=pypi_0
openssl=1.1.1t=h7f8727e_0
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.0=pypi_0
paramiko=3.1.0=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pathspec=0.11.1=pypi_0
pathy=0.10.1=pypi_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.5.0=pypi_0
pip=23.0.1=py39h06a4308_0
platformdirs=3.5.0=pyhd8ed1ab_0
preshed=3.0.8=pypi_0
prompt-toolkit=3.0.38=pyha770c72_0
prompt_toolkit=3.0.38=hd8ed1ab_0
protobuf=4.22.3=pypi_0
psutil=5.9.0=py39h5eee18b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pypi_0
pydantic=1.10.7=pypi_0
pygments=2.15.1=pyhd8ed1ab_0
pymysql=1.0.3=pypi_0
pynacl=1.5.0=pypi_0
pyparsing=3.0.9=pypi_0
python=3.9.16=h7a1cb2a_2
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytz=2023.3=pypi_0
pyyaml=6.0=pypi_0
pyzmq=19.0.2=py39hb69f2a1_2
readline=8.2=h5eee18b_0
regex=2023.3.23=pypi_0
requests=2.28.2=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.12.2=pypi_0
setuptools=67.6.1=pypi_0
six=1.16.0=pyh6c4a22f_0
smart-open=6.3.0=pypi_0
spacy=3.5.2=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.4=pypi_0
sqlite=3.41.2=h5eee18b_0
srsly=2.4.6=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
stanza=1.5.0=pypi_0
sympy=1.11.1=pypi_0
thinc=8.1.9=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=pypi_0
torch=2.0.0=pypi_0
tornado=6.1=py39hb9d737c_3
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pyhd8ed1ab_0
transformers=4.28.0=pypi_0
triton=2.0.0=pypi_0
typer=0.7.0=pypi_0
typing-extensions=4.5.0=hd8ed1ab_0
typing_extensions=4.5.0=pyha770c72_0
tzdata=2023.3=pypi_0
urllib3=1.26.15=pypi_0
wasabi=1.1.1=pypi_0
wcwidth=0.2.6=pyhd8ed1ab_0
wheel=0.40.0=pypi_0
widgetsnbextension=4.0.7=pypi_0
xz=5.2.10=h5eee18b_1
zeromq=4.3.4=h9c3ff4c_1
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
